--- controller/src/api/utils/target_template_allocation.py
@@ -1,27 +1,0 @@
-# this module should
-# determine how many templates are necessary 
-# for the subscription
-# allocate each of the emails targets to that subscription
-# return the list of targets with their email
-
-# first take the list of targets and the list of low, moderate, and high
-# next determine how many templates we should have
-# max templates gets the number of templates in the system
-# then if template usage is conservative burn,
-#    max templates is the less of 15 or #of targets divided by 3# 
-# if template usage is balanced max templates is of targets divided by 3
-# if template using is hardest to detect then utilize all templates. 
-# however there cannot be more templates than targets
-# so max templates = # of targets
-
-# first get the number of targets total
-# run the template determination to get the rough number of templates required
-# retreive the set number
-# for 
-
-# def AssignTemplatesToTargets()
-
-# def DetermineRoughTargetCount
-
-
-
