+++ controller/src/api/views/sendingprofile_views.py
@@ -3,15 +3,11 @@ Sending Profile Views.
 This handles the api for all the Sending Profile urls.
 """
 # Standard Python Libraries
-import copy
 import logging
 
 # Third-Party Libraries
 # Local
 from api.manager import CampaignManager
-from api.models.customer_models import CustomerModel, validate_customer
-from api.models.subscription_models import SubscriptionModel, validate_subscription
-from api.models.template_models import TemplateModel, validate_template
 from api.serializers.sendingprofile_serializers import (
     SendingProfileDeleteResponseSerializer,
     SendingProfileDeleteSerializer,
@@ -19,14 +15,6 @@ from api.serializers.sendingprofile_serializers import (
     SendingProfilePatchSerializer,
     SendingProfileSerializer,
 )
-from api.utils.db_utils import (
-    delete_single,
-    get_list,
-    get_single,
-    save_single,
-    update_single,
-)
-from api.utils.template_utils import format_ztime, personalize_template
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status
 from rest_framework.response import Response
@@ -127,8 +115,12 @@ class SendingProfileView(APIView):
         sp.host = self.__setAttribute(sp.host, patch_data, "host")
         sp.username = self.__setAttribute(sp.username, patch_data, "username")
         sp.password = self.__setAttribute(sp.password, patch_data, "password")
-        sp.ignore_cert_errors = self.__setAttribute(sp.ignore_cert_errors, patch_data, "ignore_cert_errors")
-        sp.from_address = self.__setAttribute(sp.from_address, patch_data, "from_address")
+        sp.ignore_cert_errors = self.__setAttribute(
+            sp.ignore_cert_errors, patch_data, "ignore_cert_errors"
+        )
+        sp.from_address = self.__setAttribute(
+            sp.from_address, patch_data, "from_address"
+        )
         sp.headers = self.__setAttribute(sp.headers, patch_data, "headers")
 
         campaign_manager.put_sending_profile(sp)
